Implemented fix by Richard: exception led to Success(Failure(..))
A Java-style try-catch construct that returned a Failure in case the catch
was executed was passed to Try.apply. Try.apply will wrap the result of the
function that is passed into it as an argument in a Success. Only if the
function that is passed into Try.apply throws an exception will Try.apply
convert it to a Failure. See the source code of Try.apply, which is simple
enough.

The net result was that a Failure was wrapped in a Success, so it was not
noticed and the program failed silently.